refactor:#29-carousel부분 코드 리팩토링
중복되는 코드 제거 및 정리
